requirements.txt
backports.weakref==1.0.post1
bleach==1.5.0
cachetools==2.0.1
funcsigs==1.0.2
google-api-python-client==1.6.4
google-auth==1.2.0
google-auth-httplib2==0.0.2
html5lib==0.9999999
httplib2==0.10.3
Markdown==2.6.9
mock==2.0.0
oauth2client==4.1.2
Paste==2.0.3
pbr==3.1.1
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
rsa==3.4.2
uritemplate==3.0.0
webapp2==2.5.2
WebOb==1.7.3
Werkzeug==0.12.2
GoogleAppEngineCloudStorageClient==1.9.22.1
